api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
numpy>=1.26.4
scipy>=1.13.1
pandas>=2.2.2
pydantic>=2.7.1
python-multipart>=0.0.9
openai>=1.30.1
python-dotenv>=1.0.1
pytest>=8.2.0
mangum>=0.17.0
# statsmodels>=0.14.0 (Removed for Vercel size limit)
# prophet>=1.1.5 (Removed for Vercel size limit)
